Refactoring charging out of the Battery class
