+++ Assets/Scripts/Environment/Hangar/BatteryCharging.cs
@@ -1,0 +1,81 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using System;
+
+public class BatteryCharging : MonoBehaviour
+{
+    public bool IsCharged { get; set; }
+
+    [SerializeField] private MeshRenderer meshRenderer;
+    private Color depletedEmission;
+    private Color chargedEmission;
+
+    private void Awake()
+    {
+        Init();
+    }
+
+    public void Init()
+    {
+        if (meshRenderer == null)
+        {
+            Debug.Log(
+                $"Getting MeshRenderer on battery {gameObject.name} as the reference is missing on the prefab.");
+
+            meshRenderer = GetComponent<MeshRenderer>();
+        }
+
+        if (meshRenderer != null)
+        {
+            meshRenderer.material.EnableKeyword("_EMISSION");
+            depletedEmission = meshRenderer.material.GetColor("_EmissionColor"); // Depleted by default
+            chargedEmission = depletedEmission * HangarConstants.BatteryEmissionCoefficient;
+        }
+        else
+        {
+            Debug.LogError($"MeshRenderer on battery {gameObject.name} is null");
+        }
+    }
+
+    public void Charge()
+    {
+        IsCharged = true;
+        SetEmission();
+    }
+
+    public void Discharge()
+    {
+        IsCharged = false;
+        SetEmission();
+    }
+
+    private void SetEmission()
+    {
+        Color emission = IsCharged ? chargedEmission : depletedEmission;
+        meshRenderer.material.SetColor("_EmissionColor", emission);
+    }
+
+    #region Persistence
+    public const string FOLDER_NAME = "HangarSaveData";
+    public const string FILE_NAME = "BatteryChargingSaveData";
+    public static string FILE_PATH
+    {
+        get => DataUtils.GetSaveFilePath(FOLDER_NAME, FILE_NAME);
+    }
+
+    public void LoadData(BatteryChargingSaveData saveData)
+    {
+        IsCharged = saveData.IsCharged;
+        SetEmission();
+    }
+    #endregion
+}
+
+[Serializable]
+public struct BatteryChargingSaveData
+{
+    public bool IsCharged;
+}
+
+
