+++ Assets/Scripts/Environment/Hangar/BatteryCharging.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d46563b4c4acac94bba8aa573acb1d78
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
