+++ Assets/Scripts/Environment/Hangar/BatteryInteractable.cs.meta
